+++ src/engine/Entity.java
@@ -15,7 +15,7 @@ public abstract class Entity implements Displayable{
 	}
 
 	@Override
-	public void update(long dt){
+	public void update(float dt){
 		this.position.x += (velocity.getX() * dt / 100.0f);
 		this.position.y += (velocity.getY() * dt / 100.0f);
 		this.position.z += (velocity.getZ() * dt / 100.0f);
